+++ flask_celery/exceptions.py
@@ -1,0 +1,5 @@
+
+class OtherInstanceError(Exception):
+    """Raised when Celery task is already running, when lock exists and has not timed out."""
+
+    pass
