+++ tests/test_timeout.py
@@ -4,7 +4,8 @@ import time
 
 import pytest
 
-from flask_celery import _LockManager, OtherInstanceError
+from flask_celery.lock_manager import LockManager
+from flask_celery.exceptions import OtherInstanceError
 from tests.instances import celery, app
 
 
@@ -16,14 +17,14 @@ def test_instances(task_name, timeout):
     """Test task instances."""
     manager_instance = list()
     task = celery.tasks[task_name]
-    original_exit = _LockManager.__exit__
+    original_exit = LockManager.__exit__
 
     def new_exit(self, *_):
         manager_instance.append(self)
         return original_exit(self, *_)
-    setattr(_LockManager, '__exit__', new_exit)
+    setattr(LockManager, '__exit__', new_exit)
     task.apply_async(args=(4, 4)).get()
-    setattr(_LockManager, '__exit__', original_exit)
+    setattr(LockManager, '__exit__', original_exit)
     assert timeout == manager_instance[0].timeout
 
 
@@ -32,12 +33,12 @@ def test_settings(key, value):
     """Test different Celery time limit settings."""
     celery.conf.update({key: value})
     manager_instance = list()
-    original_exit = _LockManager.__exit__
+    original_exit = LockManager.__exit__
 
     def new_exit(self, *_):
         manager_instance.append(self)
         return original_exit(self, *_)
-    setattr(_LockManager, '__exit__', new_exit)
+    setattr(LockManager, '__exit__', new_exit)
     tasks = [
         ('tests.instances.mul', 20), ('tests.instances.add', value), ('tests.instances.add2', 70),
         ('tests.instances.add3', 80)
@@ -47,7 +48,7 @@ def test_settings(key, value):
         task = celery.tasks[task_name]
         task.apply_async(args=(4, 4)).get()
         assert timeout == manager_instance.pop().timeout
-    setattr(_LockManager, '__exit__', original_exit)
+    setattr(LockManager, '__exit__', original_exit)
 
     celery.conf.update({key: None})
 
@@ -57,16 +58,16 @@ def test_expired():
     celery.conf.update({'CELERYD_TASK_TIME_LIMIT': 5})
     manager_instance = list()
     task = celery.tasks['tests.instances.add']
-    original_exit = _LockManager.__exit__
+    original_exit = LockManager.__exit__
 
     def new_exit(self, *_):
         manager_instance.append(self)
         return None
-    setattr(_LockManager, '__exit__', new_exit)
+    setattr(LockManager, '__exit__', new_exit)
 
     # Run the task and don't remove the lock after a successful run.
     assert 8 == task.apply_async(args=(4, 4)).get()
-    setattr(_LockManager, '__exit__', original_exit)
+    setattr(LockManager, '__exit__', original_exit)
 
     # Run again, lock is still active so this should fail.
     with pytest.raises(OtherInstanceError):
